doc-shared/azure-api-platform/src/functions/requirements.txt
azure-functions>=1.17.0
azure-identity>=1.15.0
azure-storage-blob>=12.19.0
azure-cosmos>=4.5.0
pyodbc>=5.0.0
